c_dumper: make sure ENUM values are unique
In C, ENUM values are not namespaced, so can conflict. This commit adds
a cache and resolution logic for ENUM values. For non-anonymous ENUMs,
in case of name conflict, value is rewritten as
`<enum-value>__<enum-name`, while for anonymous enums it's written as
`<enum-value>__<enum-value-version>`, similar to how
struct/union/typedef/enum names are resolved already.

Signed-off-by: Andrii Nakryiko <andriin@fb.com>