Avoid trying to mmap sysfs
Fix `btf dump /sys/kernel/btf/vmlinux` by reading the entire file into
memory when it is a BTF file. This is an imperfect heuristic for that
file being in sysfs, but it's probably good enough for now.

The lifetime management here got a bit messier, but I don't think we can
do better given that `Btf` carries its backing storage's lifetime.